assistant_backend/requirements.txt
gTTS~=2.5.4
Flask~=3.1.0
python-decouple==3.6
Flask-SQLAlchemy==3.1.0
Flask-Login==0.6.3
Flask-Migrate==3.1.0
#psycopg2==2.9.10
psycopg2-binary==2.9.10
mailjet_rest==1.3.0
google-auth
google-api-python-client~=2.153.0
google-auth-oauthlib~=1.2.1
google-auth-httplib2
pytz~=2024.2
requests
openai-whisper==20240930
SQLAlchemy~=2.0.36
Werkzeug~=3.1.3
alembic~=1.14.0
protobuf~=5.28.3
numpy~=1.26.4
Levenshtein~=0.26.1
Flask-Cors~=5.0.0
dateparser~=1.2.0